Move webDav into module
