+++ packages/service-worker/src/main.js
@@ -1,4 +1,4 @@
-import {Client} from './some.js'
+import { WebDav } from './webDav.js'
 importScripts("https://storage.googleapis.com/workbox-cdn/releases/4.3.1/workbox-sw.js");
 
 console.log('%c🔨 ServiceWorker initialized ', 'background: #273d3d; color: white');
@@ -12,34 +12,12 @@ workbox.precaching.precacheAndRoute(self.__WB_MANIFEST, {
 self.addEventListener('install', e => {
   self.skipWaiting();
 });
-class WebDav {
-  constructor() {
-    this.webDavPath = "https://host.docker.internal:9200/remote.php/dav";
-  }
-  async moveFile(sourceSpaceId, sourcePath, targetSpaceId, targetPath, token){
-    return fetch(`${this.webDavPath}/spaces/${sourceSpaceId}/${sourcePath}`, {
-      method: 'COPY',
-      mode: 'cors',
-      cache: 'no-cache',
-      credentials: 'same-origin',
-      headers: {
-        authorization: `Bearer ${token}`,
-        overwrite: 'F',
-        destination: `${this.webDavPath}/spaces/${targetSpaceId}/${targetPath}`,
-        'OCS-APIREQUEST': 'true'
-      },
-      redirect: 'follow',
-      referrerPolicy: 'no-referrer'
-    });
-  }
-}
+
 const client = new WebDav()
 addEventListener('message', async (event) => {
   if (event.data.type === 'health') {
     event.ports[0].postMessage(true);
     console.log('%c🔨 ServiceWorker up and running ', 'background: green; color: white');
-    /*var t = new Client();
-    t.test();*/
   }
   if (event.data.type === 'copy') {
     const data = event.data
