--- packages/service-worker/src/some.js
@@ -1,5 +1,0 @@
-export class Client {
-    test() {
-        console.log("LOL IM A MODULE")
-    }
-}
